requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.28.62
botocore==1.31.62
Brotli==1.0.9
cachetools==5.3.0
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.5.3
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.1
datefinder==0.7.3
dill==0.3.6
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-sslserver==0.22
django-storages==1.14.2
django-timezone-field==6.0.1
filelock==3.9.0
Flask==2.2.5
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.3.3
gevent==22.10.2
geventhttpclient==2.0.9
google-api-core==2.11.0
google-api-python-client==2.86.0
googleapis-common-protos==1.59.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kombu==5.3.2
locust==2.15.1
MarkupSafe==2.1.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
Naked==0.1.32
networkx==3.0
nltk==3.8.1
numpy==1.24.2
numpyencoder==0.3.0
oauthlib==3.2.2
openai==0.27.4
Pillow==9.4.0
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
pysqlite3==0.5.0
pysqlite3-binary==0.5.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
redis==5.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
roundrobin==0.0.4
rsa==4.9
s3transfer==0.7.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
stripe==5.4.0
sympy==1.11.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.9
Werkzeug==2.2.3
word2number==1.1
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.28.62
botocore==1.31.62
Brotli==1.0.9
cachetools==5.3.0
celery==5.3.4
certifi==2023.11.17
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.5.3
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.1
dataclasses-json==0.6.3
datefinder==0.7.3
dill==0.3.6
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-redis==5.2.0
django-sslserver==0.22
django-storages==1.14.2
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
emoji==2.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.9.0
filetype==1.2.0
Flask==2.2.5
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.3.3
fsspec==2023.4.0
gevent==22.10.2
geventhttpclient==2.0.9
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kombu==5.3.2
langchain==0.0.316
langdetect==1.0.9
langsmith==0.0.71
locust==2.15.1
lxml==4.9.3
MarkupSafe==2.1.2
marshmallow==3.20.1
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
Naked==0.1.32
networkx==3.0
nltk==3.8.1
numpy==1.24.2
numpyencoder==0.3.0
oauthlib==3.2.2
openai==0.27.4
openapi-pydantic==0.3.2
packaging==23.2
Pillow==9.4.0
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
PyJWT==1.7.1
pyparsing==3.0.9
pypdf==3.17.3
pysqlite3==0.5.0
pysqlite3-binary==0.5.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-iso639==2023.12.11
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
rapidfuzz==3.5.2
redis==5.0.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
rsa==4.9
s3transfer==0.7.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.3
stripe==5.4.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tomli==2.0.1
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2022.7
unstructured==0.11.5
unstructured-client==0.15.0
uritemplate==4.1.1
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.9
Werkzeug==2.2.3
wikipedia==1.4.0
word2number==1.1
wrapt==1.16.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
